Silence -Wconstant-conversion in the millis() rollover test
test_staleness_survives_millis_rollover deliberately adds to a value near
0xFFFFFFFF and relies on unsigned wraparound. On native, unsigned long is
64-bit, so the compiler evaluates the constant addition in 64 bits and warns
that truncating it back to uint32_t changes the value - which is exactly the
wraparound the test means to exercise. An explicit static_cast<uint32_t>
makes that intent explicit instead of relying on an implicit narrowing
conversion, with no change to what the test asserts.

+++ test/test_channel/test_channel.cpp
@@ -132,7 +132,9 @@ static void test_staleness_survives_millis_rollover(void) {
 
     // 10 s later, having wrapped through zero. Unsigned subtraction keeps the
     // elapsed time correct, so this must read as fresh rather than as 49 days.
-    const uint32_t after = before + 10000UL;
+    // The addition is deliberately made to wrap: cast explicitly so the
+    // compiler knows the truncation to uint32_t is intentional.
+    const uint32_t after = static_cast<uint32_t>(before + 10000UL);
     TEST_ASSERT_FALSE(c.isStale(after));
 
     TEST_ASSERT_TRUE(c.isStale(before + kOfficeStaleMs + 1000UL));
